+++ packages/mds-rpc-common/client/index.ts
@@ -15,10 +15,10 @@
  */
 
 import { NodeHttpTransport } from '@improbable-eng/grpc-web-node-http-transport'
+import { ClientRpcError } from '@lacuna-tech/rpc_ts/lib/client/errors'
+import { ModuleRpcProtocolClient } from '@lacuna-tech/rpc_ts/lib/protocol/client'
 import { ServiceError, ServiceResponse } from '@mds-core/mds-service-helpers'
 import { AnyFunction } from '@mds-core/mds-types'
-import { ClientRpcError } from 'rpc_ts/lib/client/errors'
-import { ModuleRpcProtocolClient } from 'rpc_ts/lib/protocol/client'
 import { RpcServiceDefinition, RPC_HOST, RPC_PORT } from '../@types'
 import { RpcCommonLogger as logger } from '../logger'