+++ container-images/mds-collector-backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @container-images/mds-collector-backend
 
+## 0.0.18
+
+### Patch Changes
+
+- 5321cca0: Upgrade to node:16.11.1-alpine
+  - @mds-core/mds-collector-backend@0.1.17
+
 ## 0.0.17
 
 ### Patch Changes