+++ container-images/mds-compliance-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @container-images/mds-compliance-service
 
+## 0.0.22
+
+### Patch Changes
+
+- 5321cca0: Upgrade to node:16.11.1-alpine
+- Updated dependencies [5c02e73f]
+  - @mds-core/mds-compliance-service@0.6.0
+
 ## 0.0.21
 
 ### Patch Changes