+++ container-images/mds-policy-author/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @container-images/mds-policy-author
 
+## 0.1.48
+
+### Patch Changes
+
+- 5321cca0: Upgrade to node:16.11.1-alpine
+  - @mds-core/mds-policy-author@0.5.7
+
 ## 0.1.47
 
 ### Patch Changes