+++ container-images/mds-policy-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @container-images/mds-policy-service
 
+## 0.0.19
+
+### Patch Changes
+
+- 5321cca0: Upgrade to node:16.11.1-alpine
+  - @mds-core/mds-policy-service@0.6.5
+
 ## 0.0.18
 
 ### Patch Changes