+++ packages/mds-api-authorizer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mds-core/mds-api-authorizer
 
+## 0.1.43
+
+### Patch Changes
+
+- Updated dependencies [0a238253]
+- Updated dependencies [0a238253]
+- Updated dependencies [0a238253]
+- Updated dependencies [5c02e73f]
+  - @mds-core/mds-types@0.6.0
+  - @mds-core/mds-utils@0.2.6
+  - @mds-core/mds-test-data@0.4.0
+
 ## 0.1.42
 
 ### Patch Changes