+++ packages/mds-geography-author/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @mds-core/mds-geography-author
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [0a238253]
+- Updated dependencies [0a238253]
+- Updated dependencies [259ad4bb]
+- Updated dependencies [5c02e73f]
+  - @mds-core/mds-types@0.6.0
+  - @mds-core/mds-utils@0.2.6
+  - @mds-core/mds-db@0.4.8
+  - @mds-core/mds-api-helpers@0.1.48
+  - @mds-core/mds-providers@0.1.41
+  - @mds-core/mds-schema-validators@0.3.8
+  - @mds-core/mds-api-server@0.3.8
+
 ## 0.2.5
 
 ### Patch Changes