+++ packages/mds-repository/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mds-core/mds-repository
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [0a238253]
+- Updated dependencies [0a238253]
+- Updated dependencies [5c02e73f]
+  - @mds-core/mds-types@0.6.0
+  - @mds-core/mds-utils@0.2.6
+  - @mds-core/mds-providers@0.1.41
+
 ## 0.1.16
 
 ### Patch Changes