+++ packages/mds-service-helpers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mds-core/mds-service-helpers
 
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [0a238253]
+- Updated dependencies [0a238253]
+- Updated dependencies [5c02e73f]
+  - @mds-core/mds-types@0.6.0
+  - @mds-core/mds-utils@0.2.6
+
 ## 0.3.9
 
 ### Patch Changes